moved BookController functions to UserController
