+++ routes/web.php
@@ -28,8 +28,8 @@ Route::middleware(['role:user'])->group(function () {
     //    if user has the role 'user', he can access these routes.
     //    put user routes here.
     Route::get('/user', [UserController::class, 'index']);
-    Route::get('/books', [BookController::class, 'showBooks']);
-    Route::get('/reserve-book/{id}', [BookController::class, 'reserveBook']);
+    Route::get('/books', [UserController::class, 'showBooks']);
+    Route::get('/reserve-book/{id}', [UserController::class, 'reserveBook']);
 
 });
 Route::middleware(['role:admin'])->group(function () {
